SW-2818 Map service + refactor
- create a MapService to fetch mapbox token and other utils
- moved types into src/types
- a lot of the code change is from import path changes
