--- src/components/Map/MapUtils.ts
@@ -1,60 +1,0 @@
-import { MapBoundingBox, MapGeometry } from './MapModels';
-
-export const getBoundingBox = (geometries: MapGeometry[]): MapBoundingBox => {
-  let llx = 0;
-  let lly = 0;
-  let urx = 0;
-  let ury = 0;
-  let first = true;
-
-  const comparePoint = (x: number, y: number) => {
-    if (first) {
-      llx = x;
-      urx = x;
-      lly = y;
-      ury = y;
-      first = false;
-      return;
-    }
-    if (x < llx) {
-      llx = x;
-    }
-    if (x > urx) {
-      urx = x;
-    }
-    if (y < lly) {
-      lly = y;
-    }
-    if (y > ury) {
-      ury = y;
-    }
-  };
-
-  const scanArray = (coord: number[]) => {
-    if (!Array.isArray(coord) || coord.length !== 2) {
-      return;
-    }
-    const [x, y] = coord;
-    if (x !== undefined && y !== undefined) {
-      comparePoint(x, y);
-    }
-  };
-
-  const scanGeometry = (geom: MapGeometry) => {
-    if (!Array.isArray(geom) || !geom.length) {
-      return;
-    }
-    (geom as number[][][][]).forEach((polygonList) => {
-      polygonList.forEach((polygon) => {
-        polygon.forEach(scanArray);
-      });
-    });
-  };
-
-  geometries.forEach(scanGeometry);
-
-  return {
-    lowerLeft: [llx, lly],
-    upperRight: [urx, ury],
-  };
-};
