+++ src/components/NurseryWithdrawals/WithdrawalDetails/WithdrawalTabPanelContent.tsx
@@ -3,7 +3,7 @@ import strings from 'src/strings';
 import Photos from './sections/Photos';
 import OutplantWithdrawalTable from './sections/OutplantWithdrawalTable';
 import { Batch, NurseryWithdrawal } from 'src/api/types/batch';
-import { Delivery } from 'src/api/types/tracking';
+import { Delivery } from 'src/types/Tracking';
 import { Species } from 'src/types/Species';
 import { WithdrawalSummary } from '../NurseryWithdrawalsDetails';
 import WithdrawalOverview from './WithdrawalOverview';
