+++ src/services/index.ts
@@ -3,6 +3,7 @@ import CachedUserService from './CachedUserService';
 import FacilityService from './FacilityService';
 import HttpService from './HttpService';
 import I18nService from './I18nService';
+import MapService from './MapService';
 import OrganizationUserService from './OrganizationUserService';
 import OrganizationService from './OrganizationService';
 import PreferencesService from './PreferencesService';
@@ -19,6 +20,7 @@ export {
   FacilityService,
   HttpService,
   I18nService,
+  MapService,
   OrganizationService,
   OrganizationUserService,
   PreferencesService,
